Move tests into right directory
